requirements.txt
streamlit 
langchain
langchain-community
langchain-openai
qdrant-client 
python-dotenv
pymongo
sentence-transformers
langchain-huggingface
